Large collection / institution / service refactoring
use singleton pattern for ServiceList InstitutionList instead of class methods. (makes testing easier).

service.id is now service.service_id to avoid conflicting with ActiveRecord "id".  Returns the unique token used to identify the service in services.yml.

institution is not in the db, only in memory, like service.

collection is built from scratch on every call, not cached in session. Nothing collection _currently_ does is expensive. If it needs to do expensive things in the future (like looking up worldcat registry), the result of those particular expensive functions can be cached in session, not entire collection.

Service objects are almost always newly instantiated when needed -- every time you ask for a Session object, you get a new one, no need to worry about sharing an object with someone else, and consequent concurrency issues. As a result, Service plug-ins can be written going forward in a much more object-oriented way, storing state in instance variables knowing that state is unique to this particular instance. 

Service objects having handle() called on them will have access to the current Umlaut Request in the request() method.  

Service objects can now read from and write to the Rails Session if needed. This took a bit of weird hacking that will have to be updated when we upgrade to Rails 2.2, which changed it's session architecture. 

There are unit tests now for collection, service, service_list, and institution_list