replace ServiceList with ServiceStore.
Is it an improvement? Barely. It's terrible we need to keep this global list of Services,
and instantiate them from ActiveRecord model objects. But it's the architecture we've inherited 9from myself), too hard to change now, can't figure out a good way to do it. Sorry future me, from past me.