HTMLUpdater, i18n, restore backwards-compat
restore backwards-compat method signature for javascript constructor.
This JS object is called by external apps for the JQuery updater api,
needs to be arguments that an external app can easily come up with
(like just the base url), and ideally shoudl be backward compat
with previous umlaut.

So we add new optional third param with locale (js args are optional
at the end), instead of previous way that changed first arg to be
a complete url with locale already in there.