Filter out service_responses from bad services
Many parts of the umlaut db store a service_id. This service_id
is expected to exist in your umalut_services.yml.  If it existed
when the service_response was created, but config was changed and
no longer exists -- this used to be a fatal error.

After this change, Umlaut will try to just ignore such a
service_response when displaying results, such responses will
be filtered out and ignored (with a warning message to log).

The Bx tests depended on the old behavior (they were testing
service_responses not properly configured), I've hackily done
the minimum to make them pass again. @scotdalton